Move logic test files tp /logic/test folder
